+++ src/decide.ts
@@ -1,4 +1,3 @@
-import { loadScript } from './utils'
 import { PostHog } from './posthog-core'
 import { Compression, DecideResponse } from './types'
 import { STORED_GROUP_PROPERTIES_KEY, STORED_PERSON_PROPERTIES_KEY } from './constants'
@@ -78,7 +77,7 @@ export class Decide {
             !!response['autocaptureExceptions'] &&
             isUndefined(exceptionAutoCaptureAddedToWindow)
         ) {
-            loadScript(this.instance.requestRouter.endpointFor('assets', '/static/exception-autocapture.js'), (err) => {
+            this.instance.requestRouter.loadScript('/static/exception-autocapture.js', (err) => {
                 if (err) {
                     return logger.error(`Could not load exception autocapture script`, err)
                 }
@@ -96,7 +95,7 @@ export class Decide {
 
                     assignableWindow[`__$$ph_site_app_${id}`] = this.instance
 
-                    loadScript(scriptUrl, (err) => {
+                    this.instance.requestRouter.loadScript(scriptUrl, (err) => {
                         if (err) {
                             logger.error(`Error while initializing PostHog app with config id ${id}`, err)
                         }
