+++ src/entrypoints/recorder-v2.ts
@@ -1,1 +1,2 @@
+// This is only here for backwards compatibility
 export * from './recorder'
