+++ src/extensions/replay/sessionrecording.ts
@@ -19,7 +19,7 @@ import { PostHog } from '../../posthog-core'
 import { DecideResponse, FlagVariant, NetworkRecordOptions, NetworkRequest, Properties } from '../../types'
 import { EventType, type eventWithTime, IncrementalSource, type listenerHandler, RecordPlugin } from '@rrweb/types'
 import Config from '../../config'
-import { timestamp, loadScript } from '../../utils'
+import { timestamp } from '../../utils'
 
 import {
     isBoolean,
@@ -541,16 +541,13 @@ export class SessionRecording {
         // If recorder.js is already loaded (if array.full.js snippet is used or posthog-js/dist/recorder is
         // imported), don't load script. Otherwise, remotely import recorder.js from cdn since it hasn't been loaded.
         if (!this.rrwebRecord) {
-            loadScript(
-                this.instance.requestRouter.endpointFor('assets', `/static/recorder.js?v=${Config.LIB_VERSION}`),
-                (err) => {
-                    if (err) {
-                        return logger.error(LOGGER_PREFIX + ` could not load recorder.js`, err)
-                    }
-
-                    this._onScriptLoaded()
+            this.instance.requestRouter.loadScript(`/static/recorder.js?v=${Config.LIB_VERSION}`, (err) => {
+                if (err) {
+                    return logger.error(LOGGER_PREFIX + ` could not load recorder.js`, err)
                 }
-            )
+
+                this._onScriptLoaded()
+            })
         } else {
             this._onScriptLoaded()
         }
